Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow>=10.3.0
numpy==1.26.4
scikit-learn==1.4.2
opencv-python==4.9.0.80
matplotlib==3.9.0
scipy==1.13.0
piexif==1.1.3
soundfile==0.12.1